refactored package structure in config
