+++ src/main/java/vaultiq/session/cache/service/VaultiqSessionManagerViaCache.java
@@ -5,8 +5,8 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.stereotype.Service;
 import vaultiq.session.cache.model.ModelType;
 import vaultiq.session.cache.service.internal.VaultiqSessionCacheService;
-import vaultiq.session.config.ConditionalOnVaultiqPersistence;
-import vaultiq.session.config.VaultiqPersistenceMode;
+import vaultiq.session.config.annotation.ConditionalOnVaultiqPersistence;
+import vaultiq.session.config.annotation.model.VaultiqPersistenceMode;
 import vaultiq.session.core.model.VaultiqSession;
 import vaultiq.session.core.VaultiqSessionManager;
