+++ src/main/java/vaultiq/session/config/annotation/model/VaultiqPersistenceMethod.java
@@ -1,4 +1,4 @@
-package vaultiq.session.config;
+package vaultiq.session.config.annotation.model;
 
 public enum VaultiqPersistenceMethod {
     USE_CACHE,
